requirements.txt
pandas==2.0.3
numpy==1.26.4
yfinance==0.2.28
python-dotenv==1.0.0
ccxt==4.1.52
plotly==5.15.0
fastapi==0.95.2
uvicorn==0.22.0
python-telegram-bot==20.3
ta-lib==0.4.24
